First working version - with hacks.
Due to the Develop:Brighton conference, we got a simple and basic
version of the app up and running asap.

This version simply creates and runs tasks to gather split data and
update the db for particular data accounts / items for UK on the 6th of
June.

Now it need to be mended to work with the actual gather queue from DEV.