RWA-193: Enqueue tasks in the splits gather queue
Task-Url: https://reflection.atlassian.net/browse/RWA-193

Updates were made throughout the codebase to get it ready for
production. This commit is still under the 0.0.1-snapshot version.

In summary we have an app which can be deployed with logging set to
daily rollover with a limit to keeping 30 days worth of logs.

The project can now be packaged using maven. 'mvn package' will create
the directory target/deploy and in it you will file all the resources
that can be edited (logging, application.properties, security
certificates, etc), run-dev.sh and run-prod.sh for starting the app.
There is also a lib directory with all the dependencies and the main jar
file references these via its Manifest file.