Multiple fixes. App executes well on Gus now.
Fixed db queries and data structure. (Split Data Fetch now hold date and
time of creation of fetch independently)
Selenium timeout increased to 60 seconds.
Fixed task service (project name and an NPE),
Exported the executor service properties to the config file.
Added a build timestamp property and printing out all config options on
- startup
Exit downloading reports if the url can't be generated.
Don't generate a url for itunes if the itemid is null.
Create the temp download directory on startup if it does not exist and
exit if can't be created.
Added a rebuild and deploy script to make compiling and rsyncing the app
to Gus easier.
Added the cloud storage read/write scope to Google Auth service.